Rename component, add in proper spots
Rename component, add in proper spots, using vaalert
