+++ src/applications/check-in/api/local-mock-api/mocks/v2/feature-toggles/index.js
@@ -4,7 +4,7 @@ const generateFeatureToggles = (toggles = {}) => {
     preCheckInEnabled = true,
     checkInExperienceTranslationDisclaimerSpanishEnabled = true,
     checkInExperienceTranslationDislaimerTagalogEnabled = true,
-    checkInExperienceTravelReimbursement = true,
+    checkInExperienceTravelReimbursement = false,
     checkInExperienceTravelLogic = false,
     checkInExperience45MinuteReminder = false,
     checkInExperienceBrowserMonitoring = false,
